requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
drf-spectacular==0.29.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
multidict==6.7.0
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
postgrest==2.24.0
propcache==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python3-openid==3.2.0
PyYAML==6.0.3
realtime==2.24.0
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.29.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
zope.interface==8.1.1
